# opentype/api/font: try to avoid extra allocations