Low hanging fruit in some performance.

The out slice was reallocated several times, we can use the length of contour points as a quick guess.

```
                   │ before.txt~ │             after.txt~              │
                   │   sec/op    │   sec/op     vs base                │
  BuildSegments-32   2.435µ ± 4%   1.367µ ± 2%  -43.87% (p=0.000 n=10)

                   │ before.txt~  │              after.txt~              │
                   │     B/op     │     B/op      vs base                │
  BuildSegments-32   3.484Ki ± 0%   1.750Ki ± 0%  -49.78% (p=0.000 n=10)

                   │ before.txt~ │             after.txt~             │
                   │  allocs/op  │ allocs/op   vs base                │
  BuildSegments-32    7.000 ± 0%   1.000 ± 0%  -85.71% (p=0.000 n=10)
```